agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-intelligence-agent"
version = "0.1.0"
requires-python = ">=3.10"
description = "Phase 2 agent layer: LangGraph + AgentCore Gateway + Bedrock + in-memory checkpointer"
readme = "README.md"

dependencies = [
    "bedrock-agentcore",
    "boto3",
    "pyyaml>=6",
    "python-dotenv>=1.0",
    "langchain",
    "langchain-mcp-adapters",
    "langchain_aws",
    "langgraph",
    "matplotlib>=3.7",
    "mcp",
    "starlette>=0.37",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

# Run with PYTHONPATH=src from agent/ so imports (graph, scoping, etc.) resolve.
